入門１０ リファクタリング
